Add files via upload
BIGS907 enables users to study a subset of Finnish 27 games wherein: the dealer forms the dealer's initial hand by dealing two playing cards
face up.
 
This text file provides a record of the user's responses to prompts for user input made by the BIGS907 software application. The record show the following. The user input a set of game rules for a blackjack-like game belonging to the Finnish 27 family of table-card games. This particular blackjack-like game was written about in the specification of U.S. Patent 9162137. On October 20th 2015, the USPTO issued U.S. Patent 9162137 to Kevin Michael McDaniel for his "Method and Apparatus for Conducting a Comparing-Card Game Belonging to a Finnish 27 Family of Table-Card Games. This particular blackjack-like game is also featured in a web application. The web application allows players to play a choice of four card game belonging to the Finnish 27 family. The web application is free to play. Visit finnish27.com for more details.